style: ret 32 ruff-fund efter opgradering til ruff 0.16.1
Lint fejlede paa main, ikke kun i PR #16. Den pinnede ruff-action
resolver nu til 0.16.1, som har regler der ikke fandtes da koden blev
skrevet: RUF046, RUF059, UP037 og skaerpet RUF012.

27 rettet automatisk (importsortering, unoedvendige noqa, citerede
typeannotationer). De sidste 5 manuelt: round() returnerer allerede int,
to mutable class attributes annoteret som ClassVar, ubrugt unpack givet
underscore, og exec-bit sat paa verify_release.py der har shebang.

Ingen adfaerdsaendring. 224 tests groenne.

+++ custom_components/intex_pool/number.py
@@ -345,10 +345,10 @@ class IntexNumber(IntexPoolEntity, NumberEntity):
     async def async_set_native_value(self, value: float) -> None:
         desc = self.entity_description
         if desc.scale is not None:
-            raw = int(round(value / desc.scale))
+            raw = round(value / desc.scale)
         else:
             raw = int(value) if float(value).is_integer() else value
         try:
             await self.coordinator.async_issue(desc.source, raw)
-        except Exception as err:  # noqa: BLE001
+        except Exception as err:
             raise HomeAssistantError(f"Failed to set {self.entity_id}: {err}") from err
